requirements.txt
fastmcp
httpx>=0.25.0
